requirements.txt
pybids==0.18.1
mne-bids==0.16.0